requirements.txt
asgiref==3.2.5
Django==3.0.4
django-debug-toolbar==2.2
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
